add annotations and constraints in airQuality and WeatherData
